feat(workflows): introduce WorkflowTemplateService for managing workflow templates
- Added WorkflowTemplateService to handle default workflows and YAML template management.
- Refactored WorkflowsService to remove template-related methods and utilize WorkflowTemplateService.
- Updated WorkflowsController to list templates using the new service.
- Modified WorkspacesService to call createDefaultWorkflows from WorkflowTemplateService during workspace creation.
- Renamed domain_discovery.yaml template and adjusted its event triggers.
- Updated unit tests to reflect changes in service dependencies and functionality.
