+++ console/src/pages/asset-group/components/tool-pipeline-builder.tsx
@@ -410,7 +410,7 @@ export function ToolPipelineBuilder({
           )}
           {custom && <Badge variant="secondary">Custom</Badge>}
           {usesDefault && <Badge variant="soft">Using default</Badge>}
-          {!ready && <Badge variant="warning">Needs config</Badge>}
+
         </div>
         <ToolProfileSelect
           toolId={tool.id}
