+++ core-api/src/modules/asset-group/asset-group.module.ts
@@ -9,7 +9,9 @@ import { JobHistory } from '../jobs-registry/entities/job-history.entity';
 import { Workflow } from '../workflows/entities/workflow.entity';
 import { Workspace } from '../workspaces/entities/workspace.entity';
 import { AssetGroupController } from './asset-group.controller';
+import { AssetGroupAssetService } from './asset-group-asset.service';
 import { AssetGroupService } from './asset-group.service';
+import { AssetGroupWorkflowService } from './asset-group-workflow.service';
 import { AssetGroupAsset } from './entities/asset-groups-assets.entity';
 import { AssetGroupWorkflow } from './entities/asset-groups-workflows.entity';
 import { AssetGroup } from './entities/asset-groups.entity';
@@ -33,7 +35,7 @@ import { AssetGroup } from './entities/asset-groups.entity';
     ConnectorsModule,
   ],
   controllers: [AssetGroupController],
-  providers: [AssetGroupService],
-  exports: [AssetGroupService],
+  providers: [AssetGroupService, AssetGroupWorkflowService, AssetGroupAssetService],
+  exports: [AssetGroupService, AssetGroupWorkflowService, AssetGroupAssetService],
 })
 export class AssetGroupModule {}
