+++ core-api/src/modules/asset-group/asset-group.service.spec.ts
@@ -5,46 +5,21 @@ import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { Test, type TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { ConnectorRegistryService } from '../connectors/connector-registry.service';
-import { Asset } from '../assets/entities/assets.entity';
-import { JobHistory } from '../jobs-registry/entities/job-history.entity';
 import { JobsRegistryService } from '../jobs-registry/jobs-registry.service';
 import { ToolsService } from '../tools/tools.service';
 import { Workflow } from '../workflows/entities/workflow.entity';
 import type { Workspace } from '../workspaces/entities/workspace.entity';
 import { WorkspaceEncryptionService } from '@/services/workspace-encryption/workspace-encryption.service';
 import { AssetGroupService } from './asset-group.service';
+import { AssetGroupWorkflowService } from './asset-group-workflow.service';
+import { AssetGroupAssetService } from './asset-group-asset.service';
 import type { CreateAssetGroupDto } from './dto/create-asset-group.dto';
-import { AssetGroupAsset } from './entities/asset-groups-assets.entity';
 import { AssetGroupWorkflow } from './entities/asset-groups-workflows.entity';
 import { AssetGroup } from './entities/asset-groups.entity';
 import type { GetAllAssetGroupsQueryDto } from './dto/get-all-asset-groups-dto.dto';
 
 describe('AssetGroupService', () => {
   let service: AssetGroupService;
-
-  const rawLastRun = {
-    id: 'jh-1',
-    workflowId: 'wf-1',
-    createdAt: new Date('2026-08-01T10:00:00Z'),
-    updatedAt: new Date('2026-08-01T10:30:00Z'),
-    totalJobs: '3',
-    status: 'COMPLETED',
-    workflowName: 'Group Workflow - group-1',
-    jobHistoryName: 'Group Workflow - group-1',
-    jobRunType: 'MANUAL',
-  };
-
-  const rawLastRun2 = {
-    id: 'jh-2',
-    workflowId: 'wf-2',
-    createdAt: new Date('2026-08-01T09:00:00Z'),
-    updatedAt: new Date('2026-08-01T09:15:00Z'),
-    totalJobs: '1',
-    status: 'FAILED',
-    workflowName: 'Group Workflow - group-2',
-    jobHistoryName: 'Group Workflow - group-2',
-    jobRunType: 'MANUAL',
-  };
 
   const mockManager = {
     findOneBy: jest.fn(),
@@ -67,16 +42,6 @@ describe('AssetGroupService', () => {
     remove: jest.fn(),
   };
 
-  const mockAssetGroupAssetRepo = {
-    find: jest.fn(),
-    create: jest.fn<
-      Record<string, unknown>,
-      [data: Record<string, unknown>]
-    >(),
-    save: jest.fn(),
-    remove: jest.fn(),
-  };
-
   const mockAssetGroupWorkflowRepo = {
     find: jest.fn(),
     createQueryBuilder: jest.fn(),
@@ -88,11 +53,6 @@ describe('AssetGroupService', () => {
       Promise<Record<string, unknown>[]>,
       [entities: Record<string, unknown>[]]
     >(),
-  };
-
-  const mockAssetRepo = {
-    findByIds: jest.fn(),
-    createQueryBuilder: jest.fn(),
   };
 
   const mockWorkflowRepo = {
@@ -108,20 +68,6 @@ describe('AssetGroupService', () => {
   const mockScanScheduleQueue = {
     add: jest.fn(),
     removeJobScheduler: jest.fn(),
-  };
-
-  // Query builder chain mock for the per-workflow latest job history lookup
-  const createMockJobHistoryBuilder = (rawRows: unknown) => {
-    const builder = {
-      leftJoin: jest.fn().mockReturnThis(),
-      where: jest.fn().mockReturnThis(),
-      select: jest.fn().mockReturnThis(),
-      distinctOn: jest.fn().mockReturnThis(),
-      orderBy: jest.fn().mockReturnThis(),
-      addOrderBy: jest.fn().mockReturnThis(),
-      getRawMany: jest.fn().mockResolvedValue(rawRows),
-    };
-    return builder;
   };
 
   // Query builder chain mock for the asset group list query
@@ -148,10 +94,6 @@ describe('AssetGroupService', () => {
     return builder;
   };
 
-  const mockJobHistoryRepo = {
-    createQueryBuilder: jest.fn(),
-  };
-
   const mockToolsService = {
     getProfileToolIds: jest.fn().mockResolvedValue(new Set()),
   };
@@ -166,25 +108,32 @@ describe('AssetGroupService', () => {
     getDEK: jest.fn(),
   };
 
+  const mockWorkflowService = {
+    addManyWorkflows: jest.fn(),
+    removeManyWorkflows: jest.fn(),
+    updateAssetGroupWorkflow: jest.fn(),
+    runGroupWorkflowScheduler: jest.fn(),
+    removeGroupWorkflowScheduler: jest.fn(),
+    getLastRunForWorkflows: jest.fn(),
+  };
+
+  const mockAssetAssetService = {
+    addManyAssets: jest.fn(),
+    removeManyAssets: jest.fn(),
+    getAssetsByAssetGroupsId: jest.fn(),
+    getAssetsNotInAssetGroup: jest.fn(),
+  };
+
   beforeEach(async () => {
     jest.clearAllMocks();
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         { provide: getRepositoryToken(AssetGroup), useValue: mockAssetGroupRepo },
         {
-          provide: getRepositoryToken(AssetGroupAsset),
-          useValue: mockAssetGroupAssetRepo,
-        },
-        {
           provide: getRepositoryToken(AssetGroupWorkflow),
           useValue: mockAssetGroupWorkflowRepo,
         },
-        { provide: getRepositoryToken(Asset), useValue: mockAssetRepo },
         { provide: getRepositoryToken(Workflow), useValue: mockWorkflowRepo },
-        {
-          provide: getRepositoryToken(JobHistory),
-          useValue: mockJobHistoryRepo,
-        },
         {
           provide: getQueueToken(BullMQName.ASSET_GROUPS_WORKFLOW_SCHEDULE),
           useValue: mockScanScheduleQueue,
@@ -193,6 +142,8 @@ describe('AssetGroupService', () => {
         { provide: JobsRegistryService, useValue: mockJobsRegistryService },
         { provide: ConnectorRegistryService, useValue: mockConnectorRegistryService },
         { provide: WorkspaceEncryptionService, useValue: mockEncryptionService },
+        { provide: AssetGroupWorkflowService, useValue: mockWorkflowService },
+        { provide: AssetGroupAssetService, useValue: mockAssetAssetService },
         AssetGroupService,
       ],
     }).compile();
@@ -255,11 +206,7 @@ describe('AssetGroupService', () => {
       expect(result.hexColor).toBe('#3b82f6');
     });
 
-    it('should add assets, create a workflow from tools and assign it with the schedule', async () => {
-      const hostAssets = [
-        { id: 'asset-1', value: 'example.com' },
-        { id: 'asset-2', value: '10.0.0.1' },
-      ];
+    it('should add assets, create a workflow from tools and delegate workflow assignment', async () => {
       const tools = [
         { id: 'tool-1', name: 'nmap' },
         { id: 'tool-2', name: 'gobuster' },
@@ -273,33 +220,24 @@ describe('AssetGroupService', () => {
       } as CreateAssetGroupDto;
 
       mockAssetGroupRepo.findOne.mockResolvedValueOnce(undefined); // name check
-      mockAssetGroupRepo.findOne.mockResolvedValue({ id: groupId }); // addManyAssets/addManyWorkflows group lookup
-      mockAssetRepo.findByIds.mockResolvedValue(hostAssets);
-      mockAssetGroupAssetRepo.find.mockResolvedValue([]);
-      mockAssetGroupAssetRepo.create.mockImplementation((data) => ({ ...data }));
+      mockAssetAssetService.addManyAssets.mockResolvedValue({
+        message: '2 assets successfully added',
+      });
       mockManager.find.mockResolvedValue(tools);
       mockWorkflowRepo.create.mockImplementation((data) => data);
       mockWorkflowRepo.save.mockResolvedValue(savedWorkflow);
-      mockWorkflowRepo.findByIds.mockResolvedValue([savedWorkflow]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
+      mockWorkflowService.addManyWorkflows.mockResolvedValue({
+        message: '1 workflows successfully added to asset group "group-1"',
+      });
 
       const result = await service.create(dto, workspaceId);
 
       expect(result.id).toBe(groupId);
 
-      // Assets added to the group
-      expect(mockAssetGroupAssetRepo.save).toHaveBeenCalledWith(
-        expect.arrayContaining([
-          expect.objectContaining({
-            assetGroup: { id: groupId },
-            asset: { id: 'asset-1' },
-          }),
-          expect.objectContaining({
-            assetGroup: { id: groupId },
-            asset: { id: 'asset-2' },
-          }),
-        ]),
+      // Assets delegated to AssetGroupAssetService
+      expect(mockAssetAssetService.addManyAssets).toHaveBeenCalledWith(
+        groupId,
+        ['asset-1', 'asset-2'],
       );
 
       // Workflow created with one job per tool
@@ -319,18 +257,11 @@ describe('AssetGroupService', () => {
         }),
       );
 
-      // Workflow assigned to the group with the passed schedule
-      expect(mockScanScheduleQueue.add).toHaveBeenCalledWith(
-        expect.any(String),
-        { id: expect.any(String) },
-        { repeat: { pattern: '0 0 * * *' } },
-      );
-      expect(mockAssetGroupWorkflowRepo.create).toHaveBeenCalledWith(
-        expect.objectContaining({
-          assetGroup: { id: groupId },
-          workflow: { id: 'workflow-1' },
-          schedule: '0 0 * * *',
-        }),
+      // Workflow assignment delegated to workflow service
+      expect(mockWorkflowService.addManyWorkflows).toHaveBeenCalledWith(
+        groupId,
+        ['workflow-1'],
+        '0 0 * * *',
       );
     });
 
@@ -346,19 +277,16 @@ describe('AssetGroupService', () => {
       mockManager.find.mockResolvedValue(tools);
       mockWorkflowRepo.create.mockImplementation((data) => data);
       mockWorkflowRepo.save.mockResolvedValue({ id: 'workflow-1' });
-      mockWorkflowRepo.findByIds.mockResolvedValue([{ id: 'workflow-1' }]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
+      mockWorkflowService.addManyWorkflows.mockResolvedValue({
+        message: '1 workflows successfully added to asset group "group-1"',
+      });
 
       await service.create(dto, workspaceId);
 
-      expect(mockScanScheduleQueue.add).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.anything(),
-        { repeat: { pattern: CronSchedule.EVERY_3_DAYS } },
-      );
-      expect(mockAssetGroupWorkflowRepo.create).toHaveBeenCalledWith(
-        expect.objectContaining({ schedule: CronSchedule.EVERY_3_DAYS }),
+      expect(mockWorkflowService.addManyWorkflows).toHaveBeenCalledWith(
+        groupId,
+        ['workflow-1'],
+        CronSchedule.EVERY_3_DAYS,
       );
     });
 
@@ -432,9 +360,9 @@ describe('AssetGroupService', () => {
       mockManager.find.mockResolvedValue([connectorTool]);
       mockWorkflowRepo.create.mockImplementation((data) => data);
       mockWorkflowRepo.save.mockResolvedValue({ id: 'workflow-1' });
-      mockWorkflowRepo.findByIds.mockResolvedValue([{ id: 'workflow-1' }]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
+      mockWorkflowService.addManyWorkflows.mockResolvedValue({
+        message: '1 workflows successfully added to asset group "group-1"',
+      });
 
       // Mock toolsService batched profile lookup to report a profile
       (mockToolsService as any).getProfileToolIds = jest
@@ -462,9 +390,9 @@ describe('AssetGroupService', () => {
       mockManager.find.mockResolvedValue([builtInTool]);
       mockWorkflowRepo.create.mockImplementation((data) => data);
       mockWorkflowRepo.save.mockResolvedValue({ id: 'workflow-1' });
-      mockWorkflowRepo.findByIds.mockResolvedValue([{ id: 'workflow-1' }]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
+      mockWorkflowService.addManyWorkflows.mockResolvedValue({
+        message: '1 workflows successfully added to asset group "group-1"',
+      });
 
       const result = await service.create(dto, workspaceId);
       expect(result.id).toBe(groupId);
@@ -483,9 +411,9 @@ describe('AssetGroupService', () => {
       mockManager.find.mockResolvedValue([connectorA, connectorB]);
       mockWorkflowRepo.create.mockImplementation((data) => data);
       mockWorkflowRepo.save.mockResolvedValue({ id: 'workflow-1' });
-      mockWorkflowRepo.findByIds.mockResolvedValue([{ id: 'workflow-1' }]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
+      mockWorkflowService.addManyWorkflows.mockResolvedValue({
+        message: '1 workflows successfully added to asset group "group-1"',
+      });
       (mockToolsService as any).getProfileToolIds = jest
         .fn()
         .mockResolvedValue(new Set(['tool-a', 'tool-b']));
@@ -501,58 +429,9 @@ describe('AssetGroupService', () => {
     });
   });
 
-  describe('addManyWorkflows', () => {
-    it('should use the provided schedule for the repeat job and the record', async () => {
-      mockAssetGroupRepo.findOne.mockResolvedValue({ id: 'group-1' });
-      mockWorkflowRepo.findByIds.mockResolvedValue([{ id: 'workflow-1' }]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
-      mockAssetGroupWorkflowRepo.create.mockImplementation((data) => ({
-        id: 'agw-1',
-        ...data,
-      }));
-
-      await service.addManyWorkflows('group-1', ['workflow-1'], '0 0 1 * *');
-
-      expect(mockScanScheduleQueue.add).toHaveBeenCalledWith(
-        expect.any(String),
-        { id: expect.any(String) },
-        { repeat: { pattern: '0 0 1 * *' } },
-      );
-      expect(mockAssetGroupWorkflowRepo.create).toHaveBeenCalledWith(
-        expect.objectContaining({ schedule: '0 0 1 * *' }),
-      );
-    });
-
-    it('should default to EVERY_3_DAYS when no schedule is passed', async () => {
-      mockAssetGroupRepo.findOne.mockResolvedValue({ id: 'group-1' });
-      mockWorkflowRepo.findByIds.mockResolvedValue([{ id: 'workflow-1' }]);
-      mockAssetGroupWorkflowRepo.find.mockResolvedValue([]);
-      mockScanScheduleQueue.add.mockResolvedValue({ repeatJobKey: 'repeat-key-1' });
-      mockAssetGroupWorkflowRepo.create.mockImplementation((data) => ({
-        id: 'agw-1',
-        ...data,
-      }));
-
-      await service.addManyWorkflows('group-1', ['workflow-1']);
-
-      expect(mockScanScheduleQueue.add).toHaveBeenCalledWith(
-        expect.any(String),
-        { id: expect.any(String) },
-        { repeat: { pattern: CronSchedule.EVERY_3_DAYS } },
-      );
-    });
-  });
-
   describe('getAssetGroupById', () => {
     const workspaceId = 'workspace-uuid';
     const groupId = 'group-1';
-
-    beforeEach(() => {
-      mockJobHistoryRepo.createQueryBuilder.mockImplementation(() =>
-        createMockJobHistoryBuilder([rawLastRun]),
-      );
-    });
 
     it('should return the asset group with its workflows embedded', async () => {
       const groupWithWorkflows = {
@@ -567,6 +446,7 @@ describe('AssetGroupService', () => {
         ],
       };
       mockAssetGroupRepo.findOne.mockResolvedValue(groupWithWorkflows);
+      mockWorkflowService.getLastRunForWorkflows.mockResolvedValue(new Map());
 
       const result = await service.getAssetGroupById(groupId, workspaceId);
 
@@ -581,6 +461,19 @@ describe('AssetGroupService', () => {
     });
 
     it('should attach the latest job history as lastRun on each workflow', async () => {
+      const rawLastRun = {
+        id: 'jh-1',
+        createdAt: new Date('2026-08-01T10:00:00Z'),
+        updatedAt: new Date('2026-08-01T10:30:00Z'),
+        totalJobs: 3,
+        status: 'COMPLETED',
+        workflowName: 'Group Workflow - group-1',
+        jobHistoryName: 'Group Workflow - group-1',
+        jobRunType: 'MANUAL',
+      };
+      const lastRunMap = new Map([['wf-1', rawLastRun]]);
+      mockWorkflowService.getLastRunForWorkflows.mockResolvedValue(lastRunMap);
+
       const groupWithWorkflows = {
         id: groupId,
         name: 'Web Servers',
@@ -596,25 +489,17 @@ describe('AssetGroupService', () => {
 
       const result = await service.getAssetGroupById(groupId, workspaceId);
 
-      expect(mockJobHistoryRepo.createQueryBuilder).toHaveBeenCalledWith(
-        'jobHistory',
-      );
-      expect(result.assetGroupWorkflows[0].lastRun).toEqual({
-        id: 'jh-1',
-        createdAt: rawLastRun.createdAt,
-        updatedAt: rawLastRun.updatedAt,
-        totalJobs: 3,
-        status: 'COMPLETED',
-        workflowName: 'Group Workflow - group-1',
-        jobHistoryName: 'Group Workflow - group-1',
-        jobRunType: 'MANUAL',
-      });
+      expect(mockWorkflowService.getLastRunForWorkflows).toHaveBeenCalledWith(['wf-1']);
+      expect(result.assetGroupWorkflows[0].lastRun).toEqual(rawLastRun);
     });
 
     it('should attach a distinct lastRun per workflow', async () => {
-      mockJobHistoryRepo.createQueryBuilder.mockImplementation(() =>
-        createMockJobHistoryBuilder([rawLastRun, rawLastRun2]),
-      );
+      const lastRunMap = new Map([
+        ['wf-1', { id: 'jh-1', status: 'COMPLETED' }],
+        ['wf-2', { id: 'jh-2', status: 'FAILED' }],
+      ]);
+      mockWorkflowService.getLastRunForWorkflows.mockResolvedValue(lastRunMap);
+
       mockAssetGroupRepo.findOne.mockResolvedValue({
         id: groupId,
         name: 'Web Servers',
@@ -646,13 +531,12 @@ describe('AssetGroupService', () => {
       const result = await service.getAssetGroupById(groupId, workspaceId);
 
       expect(result.assetGroupWorkflows).toEqual([]);
-      expect(mockJobHistoryRepo.createQueryBuilder).not.toHaveBeenCalled();
+      expect(mockWorkflowService.getLastRunForWorkflows).toHaveBeenCalledWith([]);
     });
 
     it('should set lastRun to null when no job history exists', async () => {
-      mockJobHistoryRepo.createQueryBuilder.mockImplementation(() =>
-        createMockJobHistoryBuilder([]),
-      );
+      mockWorkflowService.getLastRunForWorkflows.mockResolvedValue(new Map());
+
       mockAssetGroupRepo.findOne.mockResolvedValue({
         id: groupId,
         name: 'Web Servers',
@@ -670,16 +554,9 @@ describe('AssetGroupService', () => {
     });
 
     it('should set lastRun to null when the latest job history has no jobs (totalJobs = 0)', async () => {
-      mockJobHistoryRepo.createQueryBuilder.mockImplementation(() =>
-        createMockJobHistoryBuilder([
-          {
-            ...rawLastRun,
-            id: 'jh-empty',
-            totalJobs: '0',
-            status: 'PENDING',
-          },
-        ]),
-      );
+      // Empty map means getLastRunForWorkflows skipped the row with totalJobs=0
+      mockWorkflowService.getLastRunForWorkflows.mockResolvedValue(new Map());
+
       mockAssetGroupRepo.findOne.mockResolvedValue({
         id: groupId,
         name: 'Web Servers',
@@ -737,7 +614,6 @@ describe('AssetGroupService', () => {
       expect(result.data[0].totalAssets).toBe(3);
       expect(result.data[0].lastRunAt).toEqual(lastRunAt);
       expect(result.data[0]).not.toHaveProperty('assetGroupWorkflows');
-      expect(mockAssetGroupWorkflowRepo.createQueryBuilder).not.toHaveBeenCalled();
     });
 
     it('should attach a distinct lastRunAt per group', async () => {
@@ -783,9 +659,6 @@ describe('AssetGroupService', () => {
       const result = await service.getManyAssetGroups(listQuery, workspaceId);
 
       expect(result.data).toEqual([]);
-      expect(
-        mockAssetGroupWorkflowRepo.createQueryBuilder,
-      ).not.toHaveBeenCalled();
     });
 
     it('should sort by the lastRunAt select alias when sortBy is lastRunAt', async () => {
@@ -835,7 +708,6 @@ describe('AssetGroupService', () => {
         ],
       });
       mockAssetGroupRepo.remove.mockResolvedValue(undefined);
-      mockAssetGroupAssetRepo.remove.mockResolvedValue(undefined);
       mockWorkflowRepo.delete.mockResolvedValue(undefined);
 
       const result = await service.delete(groupId);
@@ -848,9 +720,6 @@ describe('AssetGroupService', () => {
         'repeat-key-2',
       );
       expect(mockWorkflowRepo.delete).toHaveBeenCalledWith(['wf-1', 'wf-2']);
-      expect(mockAssetGroupAssetRepo.remove).toHaveBeenCalledWith([
-        { id: 'aga-1' },
-      ]);
       expect(mockAssetGroupRepo.remove).toHaveBeenCalledWith(
         expect.objectContaining({ id: groupId }),
       );
@@ -880,101 +749,6 @@ describe('AssetGroupService', () => {
       expect(mockScanScheduleQueue.removeJobScheduler).not.toHaveBeenCalled();
       expect(mockWorkflowRepo.delete).not.toHaveBeenCalled();
       expect(mockAssetGroupRepo.remove).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('getAssetsByAssetGroupsId', () => {
-    const workspaceId = 'workspace-uuid';
-    const groupId = 'group-1';
-
-    const createMockAssetSearchBuilder = ({
-      data = [] as Record<string, unknown>[],
-      total = 0,
-    } = {}) => {
-      const builder = {
-        innerJoin: jest.fn().mockReturnThis(),
-        where: jest.fn().mockReturnThis(),
-        andWhere: jest.fn().mockReturnThis(),
-        orderBy: jest.fn().mockReturnThis(),
-        skip: jest.fn().mockReturnThis(),
-        take: jest.fn().mockReturnThis(),
-        getManyAndCount: jest.fn().mockResolvedValue([data, total]),
-      };
-      return builder;
-    };
-
-    it('should filter assets by value ILIKE when search is provided', async () => {
-      mockAssetGroupRepo.findOne.mockResolvedValue({ id: groupId });
-      const builder = createMockAssetSearchBuilder({
-        data: [{ id: 'asset-1', value: 'api.example.com' }],
-        total: 1,
-      });
-      mockAssetRepo.createQueryBuilder = jest.fn().mockReturnValue(builder);
-
-      const result = await service.getAssetsByAssetGroupsId(
-        groupId,
-        {
-          page: 1,
-          limit: 10,
-          sortBy: 'createdAt',
-          sortOrder: SortOrder.DESC,
-          search: 'example',
-        },
-        workspaceId,
-      );
-
-      expect(builder.andWhere).toHaveBeenCalledWith(
-        'asset.value ILIKE :search',
-        { search: '%example%' },
-      );
-      expect(result.data).toEqual([{ id: 'asset-1', value: 'api.example.com' }]);
-      expect(result.total).toBe(1);
-    });
-
-    it('should not add the search filter when search is absent', async () => {
-      mockAssetGroupRepo.findOne.mockResolvedValue({ id: groupId });
-      const builder = createMockAssetSearchBuilder();
-      mockAssetRepo.createQueryBuilder = jest.fn().mockReturnValue(builder);
-
-      await service.getAssetsByAssetGroupsId(
-        groupId,
-        { page: 1, limit: 10, sortBy: 'createdAt', sortOrder: SortOrder.DESC },
-        workspaceId,
-      );
-
-      expect(builder.andWhere).not.toHaveBeenCalled();
-    });
-
-    it('should throw NotFoundException when the group does not belong to the workspace', async () => {
-      mockAssetGroupRepo.findOne.mockResolvedValue(undefined);
-
-      await expect(
-        service.getAssetsByAssetGroupsId(
-          groupId,
-          { page: 1, limit: 10, sortBy: 'createdAt', sortOrder: SortOrder.DESC },
-          workspaceId,
-        ),
-      ).rejects.toThrow(NotFoundException);
-      expect(mockAssetRepo.createQueryBuilder).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('removeGroupWorkflowScheduler', () => {
-    it('should remove the BullMQ scheduler for the given repeat key', async () => {
-      mockScanScheduleQueue.removeJobScheduler.mockResolvedValue(undefined);
-
-      await service.removeGroupWorkflowScheduler('repeat:agw-1:1');
-
-      expect(mockScanScheduleQueue.removeJobScheduler).toHaveBeenCalledWith(
-        'repeat:agw-1:1',
-      );
-    });
-
-    it('should do nothing when no repeat key is provided', async () => {
-      await service.removeGroupWorkflowScheduler(undefined);
-      await service.removeGroupWorkflowScheduler(null);
-
-      expect(mockScanScheduleQueue.removeJobScheduler).not.toHaveBeenCalled();
     });
   });
 });
