+++ core-api/src/modules/jobs-registry/jobs-registry.module.ts
@@ -6,6 +6,7 @@ import { AssetGroupWorkflow } from '../asset-group/entities/asset-groups-workflo
 import { IssueComment } from '../issues/entities/issue-comment.entity';
 import { Issue } from '../issues/entities/issue.entity';
 import { ConnectorsModule } from '../connectors/connectors.module';
+import { ToolConfigProfile } from '../tools/entities/tool-config-profiles.entity';
 import { WorkspacesModule } from '../workspaces/workspaces.module';
 import { JobErrorLog } from './entities/job-error-log.entity';
 import { JobHistory } from './entities/job-history.entity';
@@ -29,6 +30,7 @@ import {
       JobErrorLog,
       Issue,
       IssueComment,
+      ToolConfigProfile,
     ]),
     ConnectorsModule,
     WorkspacesModule,
