+++ core-api/src/modules/jobs-registry/processors/scan-schedule.processor.spec.ts
@@ -1,12 +1,12 @@
 import { JobRunType } from '@/common/enums/enum';
-import type { AssetGroupService } from '@/modules/asset-group/asset-group.service';
+import type { AssetGroupWorkflowService } from '@/modules/asset-group/asset-group-workflow.service';
 import type { AssetGroupWorkflow } from '@/modules/asset-group/entities/asset-groups-workflows.entity';
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import type { Job } from 'bullmq';
 import { AssetGroupsScheduleConsumer } from './scan-schedule.processor';
 
 describe('AssetGroupsScheduleConsumer', () => {
-  const mockAssetGroupService = {
+  const mockAssetGroupWorkflowService = {
     runGroupWorkflowScheduler: jest.fn(),
     removeGroupWorkflowScheduler: jest.fn(),
   };
@@ -26,7 +26,7 @@ describe('AssetGroupsScheduleConsumer', () => {
   beforeEach(() => {
     jest.clearAllMocks();
     consumer = new AssetGroupsScheduleConsumer(
-      mockAssetGroupService as unknown as AssetGroupService,
+      mockAssetGroupWorkflowService as unknown as AssetGroupWorkflowService,
     );
   });
 
@@ -35,28 +35,28 @@ describe('AssetGroupsScheduleConsumer', () => {
 
     await consumer.process(job);
 
-    expect(mockAssetGroupService.runGroupWorkflowScheduler).toHaveBeenCalledWith(
+    expect(mockAssetGroupWorkflowService.runGroupWorkflowScheduler).toHaveBeenCalledWith(
       'agw-1',
       JobRunType.SCHEDULED,
     );
-    expect(mockAssetGroupService.removeGroupWorkflowScheduler).not.toHaveBeenCalled();
+    expect(mockAssetGroupWorkflowService.removeGroupWorkflowScheduler).not.toHaveBeenCalled();
   });
 
   it('removes the BullMQ scheduler when the asset group workflow is not found', async () => {
-    mockAssetGroupService.runGroupWorkflowScheduler.mockRejectedValueOnce(
+    mockAssetGroupWorkflowService.runGroupWorkflowScheduler.mockRejectedValueOnce(
       new NotFoundException('Asset group workflow with ID "agw-1" not found'),
     );
     const job = createMockJob('agw-1', 'repeat:agw-1:1');
 
     await expect(consumer.process(job)).resolves.toBeUndefined();
 
-    expect(mockAssetGroupService.removeGroupWorkflowScheduler).toHaveBeenCalledWith(
+    expect(mockAssetGroupWorkflowService.removeGroupWorkflowScheduler).toHaveBeenCalledWith(
       'repeat:agw-1:1',
     );
   });
 
   it('removes the current job when not found and no repeat key exists', async () => {
-    mockAssetGroupService.runGroupWorkflowScheduler.mockRejectedValueOnce(
+    mockAssetGroupWorkflowService.runGroupWorkflowScheduler.mockRejectedValueOnce(
       new NotFoundException('Asset group workflow with ID "agw-1" not found'),
     );
     const job = createMockJob('agw-1', null);
@@ -64,17 +64,17 @@ describe('AssetGroupsScheduleConsumer', () => {
     await expect(consumer.process(job)).resolves.toBeUndefined();
 
     expect(job.remove).toHaveBeenCalled();
-    expect(mockAssetGroupService.removeGroupWorkflowScheduler).not.toHaveBeenCalled();
+    expect(mockAssetGroupWorkflowService.removeGroupWorkflowScheduler).not.toHaveBeenCalled();
   });
 
   it('rethrows non-not-found errors and does not remove the job', async () => {
     const error = new BadRequestException('Asset group workflow has no assets');
-    mockAssetGroupService.runGroupWorkflowScheduler.mockRejectedValueOnce(error);
+    mockAssetGroupWorkflowService.runGroupWorkflowScheduler.mockRejectedValueOnce(error);
     const job = createMockJob('agw-1', 'repeat:agw-1:1');
 
     await expect(consumer.process(job)).rejects.toBe(error);
 
-    expect(mockAssetGroupService.removeGroupWorkflowScheduler).not.toHaveBeenCalled();
+    expect(mockAssetGroupWorkflowService.removeGroupWorkflowScheduler).not.toHaveBeenCalled();
     expect(job.remove).not.toHaveBeenCalled();
   });
 });
