+++ core-api/src/modules/workflows/templates/domain_discovery.yaml
@@ -1,4 +1,4 @@
-name: vulnerability_scan_basic
+name: domain_discovery
 
 on:
   target: [domain.create, domain.re-scan]
