+++ core-api/src/modules/workflows/workflows.service.ts
@@ -1,185 +1,23 @@
 import { GetManyBaseResponseDto } from '@/common/dtos/get-many-base.dto';
-import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import * as fs from 'fs';
-import * as yaml from 'js-yaml';
-import * as path from 'path';
 import { Repository } from 'typeorm';
+import { Workflow } from './entities/workflow.entity';
 import { User } from '../auth/entities/user.entity';
 import { Workspace } from '../workspaces/entities/workspace.entity';
 import { CreateWorkflowDto } from './dto/create-workflow.dto';
 import { GetManyWorkflowsQueryDto } from './dto/get-many-workflows.dto';
-import { Workflow } from './entities/workflow.entity';
+
+const ALLOWED_WORKFLOW_SORT_FIELDS = ['createdAt', 'updatedAt', 'name'] as const;
 
 @Injectable()
-export class WorkflowsService implements OnModuleInit {
+export class WorkflowsService {
   constructor(
     @InjectRepository(Workflow)
     public workflowRepository: Repository<Workflow>,
-    @InjectRepository(Workspace)
-    private workspaceRepository: Repository<Workspace>,
   ) {}
 
   private readonly logger = new Logger(WorkflowsService.name);
-  private readonly templatesPath = path.join(__dirname, 'templates');
-
-  /**
-   * Get a template by name
-   * @param name Name of the template
-   * @returns Workflow object containing the template
-   * @throws Error if template is not found
-   */
-  public async getTemplate(name: string): Promise<Workflow> {
-    const template = await this.workflowRepository.findOne({
-      where: { name },
-    });
-
-    if (!template) {
-      throw new Error(`Template ${name} not found`);
-    }
-
-    return template;
-  }
-
-  /**
-   * List all YAML template files in the templates directory
-   * @returns Array of YAML file names
-   */
-  public listTemplates(): string[] {
-    if (!fs.existsSync(this.templatesPath)) {
-      return [];
-    }
-
-    const files = fs.readdirSync(this.templatesPath);
-
-    const yamlFiles = files.filter(
-      (file) => file.endsWith('.yaml') || file.endsWith('.yml'),
-    );
-
-    return yamlFiles;
-  }
-
-  /**
-   * Normalize the on property of the workflow
-   * @param obj Workflow object
-   * @returns Normalized workflow object
-   */
-
-  private normalizeOn(obj: Record<string, unknown>): Record<string, unknown> {
-    if (!obj.on) return obj;
-
-    for (const key of Object.keys(obj.on)) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const value = obj.on[key];
-      if (Array.isArray(value)) {
-        obj.on[key] = value.map(String);
-      } else if (typeof value === 'string') {
-        obj.on[key] = [value];
-      } else {
-        throw new Error(
-          `Invalid type for on.${key}, must be string or array of string`,
-        );
-      }
-    }
-    return obj;
-  }
-
-  async onModuleInit() {
-    try {
-      const workspaces = await this.workspaceRepository.find();
-
-      for (const workspace of workspaces) {
-        await this.createDefaultWorkflows(workspace.id);
-      }
-    } catch (error) {
-      this.logger.error('Error initializing workflows:', error);
-    }
-  }
-
-  /**
-   * Create or update default workflows for a specific workspace
-   * @param workspaceId ID of the workspace
-   */
-  public async createDefaultWorkflows(workspaceId: string) {
-    try {
-      const yamlFiles = this.listTemplates();
-
-      for (const fileName of yamlFiles) {
-        try {
-          const filePath = path.join(this.templatesPath, fileName);
-          const fileContent = fs.readFileSync(filePath, 'utf8');
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
-          const parsed = yaml.load(fileContent) as Record<string, unknown>;
-
-          const newContent = this.normalizeOn(parsed);
-
-          const baseName = fileName.replace(/\.(yaml|yml)$/, '');
-          const normalizedName = baseName
-            .replace(/[-_]/g, ' ')
-            .replace(/\s+/g, ' ')
-            .trim();
-          const workflowName =
-            normalizedName.charAt(0).toUpperCase() + normalizedName.slice(1);
-
-          // Check current workflow for this workspace
-          const existing = await this.workflowRepository.findOne({
-            where: {
-              filePath: fileName,
-              workspace: { id: workspaceId },
-            },
-          });
-
-          if (!existing) {
-            // Insert new workflow for this workspace
-            await this.workflowRepository.insert({
-              name: workflowName,
-              content: newContent,
-              filePath: fileName,
-              workspace: { id: workspaceId },
-              isCanDelete: false,
-              isCanEdit: false,
-            });
-          } else if (
-            JSON.stringify(newContent) !== JSON.stringify(existing.content) ||
-            existing.name !== workflowName
-          ) {
-            await this.workflowRepository.update(
-              { id: existing.id },
-              { content: newContent, name: workflowName },
-            );
-          }
-        } catch (error) {
-          this.logger.error(
-            `Error processing workflow ${fileName} for workspace ${workspaceId}: ${(error as Error).message}`,
-            (error as Error).stack,
-          );
-        }
-      }
-    } catch (error) {
-      this.logger.error(
-        `Error creating default workflows for workspace ${workspaceId}:`,
-        error,
-      );
-    }
-  }
-
-  /**
-   * Parse a YAML file to object
-   * @param fileName Name of the YAML file
-   * @returns Parsed YAML object
-   */
-  public parseTemplate(fileName: string): unknown {
-    const filePath = path.join(this.templatesPath, fileName);
-
-    if (!fs.existsSync(filePath)) {
-      throw new Error(`Template file ${fileName} not found`);
-    }
-
-    const fileContent = fs.readFileSync(filePath, 'utf8');
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
-    const parsed = yaml.load(fileContent);
-    return parsed as unknown;
-  }
 
   /**
    * Creates a new workflow in the specified workspace
@@ -294,8 +132,11 @@ export class WorkflowsService implements OnModuleInit {
       queryBuilder.andWhere('workflow.name LIKE :name', { name: `%${name}%` });
     }
 
-    // Apply sorting
-    queryBuilder.orderBy(`workflow.${sortBy}`, sortOrder);
+    // Apply sorting with whitelist to prevent SQL injection
+    const safeSortBy = (ALLOWED_WORKFLOW_SORT_FIELDS as readonly string[]).includes(sortBy)
+      ? sortBy
+      : 'createdAt';
+    queryBuilder.orderBy(`workflow.${safeSortBy}`, sortOrder);
 
     // Apply pagination
     const skip = (page - 1) * limit;
@@ -315,21 +156,5 @@ export class WorkflowsService implements OnModuleInit {
       pageCount,
       hasNextPage,
     };
-  }
-
-  /**
-   * Retrieves all workflows within a specific workspace
-   * @param workspace The workspace to retrieve workflows from
-   * @returns Array of Workflow entities in the workspace
-   */
-  async getWorkflowsByWorkspace(workspace: {
-    id: string;
-  }): Promise<Workflow[]> {
-    return await this.workflowRepository.find({
-      where: {
-        workspace: { id: workspace.id },
-      },
-      relations: ['createdBy', 'workspace'],
-    });
   }
 }
